@@ -32,48 +32,48 @@
 	"main": "./dist/electron/main.js",
 	"private": true,
 	"dependencies": {
-		"@angular/animations": "^8.2.14",
-		"@angular/common": "^8.2.14",
-		"@angular/compiler": "^8.2.14",
-		"@angular/core": "^8.2.14",
-		"@angular/forms": "^8.2.14",
-		"@angular/platform-browser": "^8.2.14",
-		"@angular/platform-browser-dynamic": "^8.2.14",
-		"@angular/router": "^8.2.14",
+		"@angular/animations": "^9.0.0",
+		"@angular/common": "^9.0.0",
+		"@angular/compiler": "^9.0.0",
+		"@angular/core": "^9.0.0",
+		"@angular/forms": "^9.0.0",
+		"@angular/platform-browser": "^9.0.0",
+		"@angular/platform-browser-dynamic": "^9.0.0",
+		"@angular/router": "^9.0.0",
 		"@logigator/logigator-shared-comps": "latest",
-		"@ngx-translate/core": "^11.0.1",
+		"@ngx-translate/core": "^12.0.0",
 		"bootstrap": "^4.4.1",
 		"express": "^4.17.1",
 		"file-saver": "^2.0.2",
 		"@logigator/logigator-simulation": "github:logigator/logigator-simulation",
 		"github-markdown-css": "^3.0.1",
-		"ngx-electron": "^2.1.1",
-		"ngx-toastr": "^11.2.1",
+		"ngx-electron": "^2.2.0",
+		"ngx-toastr": "^12.0.0",
 		"node-fetch": "^2.6.0",
-		"pixi.js": "^5.2.0",
+		"pixi.js": "^5.2.1",
 		"run-script-os": "^1.0.7",
-		"rxjs": "~6.4.0",
+		"rxjs": "~6.5.4",
 		"tslib": "^1.10.0",
-		"zone.js": "~0.9.1"
+		"zone.js": "~0.10.2"
 	},
 	"devDependencies": {
 		"@angular-builders/custom-webpack": "^8.4.1",
-		"@angular-devkit/build-angular": "^0.803.22",
-		"@angular/cli": "^8.3.20",
-		"@angular/compiler-cli": "^8.2.14",
-		"@angular/language-service": "^8.2.14",
-		"@ngtools/webpack": "^8.3.20",
+		"@angular-devkit/build-angular": "^0.900.1",
+		"@angular/cli": "^9.0.1",
+		"@angular/compiler-cli": "^9.0.0",
+		"@angular/language-service": "^9.0.0",
+		"@ngtools/webpack": "^9.0.1",
 		"@types/express": "^4.17.2",
 		"@types/file-saver": "^2.0.1",
 		"@types/gtag.js": "0.0.3",
 		"@types/jasmine": "~3.3.8",
 		"@types/jasminewd2": "^2.0.8",
-		"@types/node": "~8.9.4",
+		"@types/node": "^13.7.0",
 		"@types/node-fetch": "^2.5.4",
-		"codelyzer": "^5.2.0",
-		"electron": "^6.1.7",
-		"electron-packager": "^14.1.1",
-		"electron-rebuild": "^1.8.8",
+		"codelyzer": "^5.2.1",
+		"electron": "^8.0.0",
+		"electron-packager": "^14.2.1",
+		"electron-rebuild": "^1.10.0",
 		"electron-reload": "^1.5.0",
 		"env-cmd": "^10.0.1",
 		"html-loader": "^0.5.5",
@@ -87,9 +87,9 @@
 		"markdown-loader": "^5.1.0",
 		"npm-run-all": "^4.1.5",
 		"protractor": "~5.4.0",
-		"tslint": "~5.15.0",
-		"typescript": "~3.5.3",
-		"wait-on": "^3.3.0",
+		"tslint": "^6.0.0",
+		"typescript": "^3.7.5",
+		"wait-on": "^4.0.0",
 		"webpack-bundle-analyzer": "^3.6.0",
 		"webpack-preprocessor-loader": "github:logigator/preprocessor-loader"
 	},
