@@ -6,7 +6,7 @@ describe('ElementProviderService', () => {
 	beforeEach(() => TestBed.configureTestingModule({}));
 
 	it('should be created', () => {
-		const service: ElementProviderService = TestBed.get(ElementProviderService);
+		const service: ElementProviderService = TestBed.inject(ElementProviderService);
 		expect(service).toBeTruthy();
 	});
 });
