@@ -6,7 +6,7 @@ describe('ProjectInteractionService', () => {
 	beforeEach(() => TestBed.configureTestingModule({}));
 
 	it('should be created', () => {
-		const service: ProjectInteractionService = TestBed.get(ProjectInteractionService);
+		const service: ProjectInteractionService = TestBed.inject(ProjectInteractionService);
 		expect(service).toBeTruthy();
 	});
 });
