@@ -6,7 +6,7 @@ describe('ProjectSaveManagementService', () => {
 	beforeEach(() => TestBed.configureTestingModule({}));
 
 	it('should be created', () => {
-		const service: ProjectSaveManagementService = TestBed.get(ProjectSaveManagementService);
+		const service: ProjectSaveManagementService = TestBed.inject(ProjectSaveManagementService);
 		expect(service).toBeTruthy();
 	});
 });
