@@ -6,7 +6,7 @@ describe('ProjectsService', () => {
 	beforeEach(() => TestBed.configureTestingModule({}));
 
 	it('should be created', () => {
-		const service: ProjectsService = TestBed.get(ProjectsService);
+		const service: ProjectsService = TestBed.inject(ProjectsService);
 		expect(service).toBeTruthy();
 	});
 });
