@@ -6,7 +6,7 @@ describe('WorkerCommunicationNodeService', () => {
 	beforeEach(() => TestBed.configureTestingModule({}));
 
 	it('should be created', () => {
-		const service: WorkerCommunicationNodeService = TestBed.get(WorkerCommunicationNodeService);
+		const service: WorkerCommunicationNodeService = TestBed.inject(WorkerCommunicationNodeService);
 		expect(service).toBeTruthy();
 	});
 });
