@@ -6,7 +6,7 @@ describe('WorkerCommunicationService', () => {
 	beforeEach(() => TestBed.configureTestingModule({}));
 
 	it('should be created', () => {
-		const service: WorkerCommunicationWasmService = TestBed.get(WorkerCommunicationWasmService);
+		const service: WorkerCommunicationWasmService = TestBed.inject(WorkerCommunicationWasmService);
 		expect(service).toBeTruthy();
 	});
 });
