@@ -6,7 +6,7 @@ describe('ThemingService', () => {
 	beforeEach(() => TestBed.configureTestingModule({}));
 
 	it('should be created', () => {
-		const service: ThemingService = TestBed.get(ThemingService);
+		const service: ThemingService = TestBed.inject(ThemingService);
 		expect(service).toBeTruthy();
 	});
 });
