@@ -2,11 +2,12 @@
 // `ng build --prod` replaces `environment.ts` with `environment.prod.ts`.
 // The list of file replacements can be found in `angular.json`.
 
-export declare const environment: {
-	production: boolean,
-	chunkSize: number,
-	gridPixelWidth: number,
-	apiPrefix: string,
-	domain: string
+export const environment = {
+	production: false,
+	chunkSize: 16,
+	gridPixelWidth: 16,
+	apiPrefix: 'http://api.logigator-local-dev.com',
+	domain: 'http://editor.logigator-local-dev.com'
 };
-export declare const ELECTRON: boolean;
+export const ELECTRON = false;
+
